py/pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=58.1.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redacted-py"
version = "1.0.4"
description = "Redacting classified documents"
readme = "README.md"
authors = [{ name = "Cyril Dever", email = "cdever@pep-s.com" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = [
    "data",
    "obfuscation",
    "data masking",
    "redacted",
    "classified",
]
dependencies = [
    "feistel-py >= 0.2.0",
]
requires-python = ">=3.10.2"

[project.urls]
Homepage = "https://github.com/cyrildever/redacted"

[project.scripts]
redacted-py = "redacted.__main__:main"
